split instruments.ts file
